requirements.txt
fastapi
uvicorn
mangum
pydantic
python-dotenv
langchain
langchain-core
langchain-openai
langchain-groq
langchain-community
streamlit
requests